+++ app/src/main/res/layout/list_item.xml
@@ -19,7 +19,7 @@
 
         <Button
             style="?android:attr/buttonStyleSmall"
-            android:layout_width="wrap_content"
+            android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:text="Like!"
             android:textStyle="bold"
